Communication with server side is implemented
git-svn-id: https://vocrecaptor.googlecode.com/svn/trunk@54 31e652d2-c5d7-11dd-927d-974589228060

+++ vocrecaptorexercises/src/com/vocrecaptor/fx/model/bean/DictionaryBean.fx
@@ -10,7 +10,7 @@ public class DictionaryBean {
     /**
      * Dictionary id.
      */
-    public var id: Integer;
+    public var id: Long;
 
     /**
      * Dictionary description.
